move stoich into src dir
